Move sidebar ui state setting to page component
