--- src/components/sidebar/initial-state.js
@@ -1,3 +1,0 @@
-export default {
-  isPaddingActive: true
-}
