+++ src/components/sidebar/sidebar.js
@@ -2,7 +2,6 @@ import React from 'react';
 import { connect } from 'react-redux';
 import Component from './sidebar-component';
 
-import mapStateToProps from './sidebar-selectors';
 import * as actions from './sidebar-actions';
 
 const SideBarContainer = props => {
@@ -10,4 +9,4 @@ const SideBarContainer = props => {
   return <Component handleSidebarToggle={handleSidebarToggle} {...props} />; 
 }
 
-export default connect(mapStateToProps, actions)(SideBarContainer);
+export default connect(null, actions)(SideBarContainer);
