+++ src/pages/data-globe/data-globe-component.jsx
@@ -7,11 +7,11 @@ import sceneSettings from './data-globe-settings.js';
 
 const { REACT_APP_DATA_GLOBE_SCENE_ID: SCENE_ID } = process.env;
 
-const DataGlobeComponent = ({ sceneLayers, activeLayers, isLandscapeMode, handleZoomChange, handleLayerToggle }) => (
+const DataGlobeComponent = ({ sceneLayers, activeLayers, isLandscapeMode, isPaddingActive, handleZoomChange, handleLayerToggle }) => (
   <Globe sceneId={SCENE_ID} sceneSettings={sceneSettings} >
     <ArcgisLayerManager activeLayers={activeLayers}/>
     <LandscapeViewManager zoomLevelTrigger={10} onZoomChange={handleZoomChange} isLandscapeMode={isLandscapeMode}/>
-    <Sidebar>
+    <Sidebar isPaddingActive={isPaddingActive}>
       {
         sceneLayers &&
         sceneLayers.map(l => (
