+++ src/pages/data-globe/data-globe-initial-state.js
@@ -1,0 +1,6 @@
+export default {
+  globe: {},
+  ui: {
+    isPaddingActive: false
+  }
+}
