+++ src/pages/featured-globe/featured-globe-component.jsx
@@ -14,14 +14,14 @@ import ZoomWidget from 'components/widgets/zoom-widget';
 
 const { REACT_APP_FEATURED_GLOBE_SCENE_ID: SCENE_ID } = process.env;
 
-const FeaturedGlobeComponent = ({ sceneLayers, activeLayers, isLandscapeMode, handleLayerToggle, handleZoomChange }) => (
+const FeaturedGlobeComponent = ({ sceneLayers, activeLayers, isLandscapeMode, isPaddingActive, handleLayerToggle, handleZoomChange }) => (
   <Globe sceneId={SCENE_ID} sceneSettings={sceneSettings} >
     <ArcgisLayerManager activeLayers={activeLayers}/>
     <LandscapeViewManager zoomLevelTrigger={10} onZoomChange={handleZoomChange} isLandscapeMode={isLandscapeMode}/>
     <LocationWidget />
     <ZoomWidget />
     <NavigationToggleWidget />
-    <Sidebar>
+    <Sidebar isPaddingActive={isPaddingActive}>
       {
         sceneLayers &&
         sceneLayers.map(l => (
