+++ src/pages/featured-globe/featured-globe-initial-state.js
@@ -1,0 +1,8 @@
+export default {
+  globe: {
+    activeLayers: []
+  },
+  ui: {
+    isPaddingActive: true
+  }
+}
