+++ src/pages/featured-globe/featured-globe-selectors.js
@@ -1,21 +1,29 @@
 import { createSelector, createStructuredSelector } from 'reselect';
 import { getFeaturedGlobeLayers } from 'selectors/layers-selectors';
-import { selectGlobeUrlState } from 'selectors/location-selectors';
-import initialState from './initial-state';
+import { selectGlobeUrlState, selectUiUrlState } from 'selectors/location-selectors';
+import initialState from './featured-globe-initial-state';
 
 const getGlobeSettings = createSelector(selectGlobeUrlState, globeUrlState => {
   return {
-    ...initialState,
+    ...initialState.globe,
     ...globeUrlState
+  }
+})
+
+const getUiSettings = createSelector(selectUiUrlState, uiUrlState => {
+  return {
+    ...initialState.ui,
+    ...uiUrlState
   }
 })
 
 const getActiveLayers = createSelector(getGlobeSettings, globeSettings => globeSettings.activeLayers)
 const getLandscapeMode = createSelector(getGlobeSettings, globeSettings => globeSettings.landscapeView)
+const getScenePadding = createSelector(getUiSettings, uiSettings => uiSettings.isPaddingActive)
 
 export default createStructuredSelector({
   sceneLayers: getFeaturedGlobeLayers,
   activeLayers: getActiveLayers,
-  globeSettings: getGlobeSettings,
-  isLandscapeMode: getLandscapeMode
+  isLandscapeMode: getLandscapeMode,
+  isPaddingActive: getScenePadding
 })
