--- src/pages/featured-globe/initial-state.js
@@ -1,3 +1,0 @@
-export default {
-  activeLayers: []
-}
